fix: move reusable components into components folder
